REF: Move `validation` module up one level
